+++ geotrek/api/v2/views/trekking.py
@@ -18,7 +18,8 @@ class TrekViewSet(api_viewsets.GeotrekViewset):
         .annotate(geom2d_transformed=Transform(F('geom'), settings.API_SRID),
                   geom3d_transformed=Transform(F('geom_3d'), settings.API_SRID),
                   length_2d_m=Length('geom'),
-                  length_3d_m=Length3D('geom_3d'))
+                  length_3d_m=Length3D('geom_3d')) \
+        .order_by('pk')  # Required for reliable pagination
     filter_fields = ('difficulty', 'themes', 'networks', 'practice')
 
     @decorators.list_route(methods=['get'])
@@ -119,7 +120,8 @@ class POIViewSet(api_viewsets.GeotrekViewset):
         .select_related('topo_object', 'type', ) \
         .prefetch_related('topo_object__aggregations', 'attachments') \
         .annotate(geom2d_transformed=Transform(F('geom'), settings.API_SRID),
-                  geom3d_transformed=Transform(F('geom_3d'), settings.API_SRID))
+                  geom3d_transformed=Transform(F('geom_3d'), settings.API_SRID)) \
+        .order_by('pk')  # Required for reliable pagination
     filter_fields = ('type',)
 
     @decorators.list_route(methods=['get'])