+++ geotrek/common/tests/__init__.py
@@ -7,6 +7,7 @@ from geotrek.common.models import FileType  # NOQA
 
 from mapentity.tests import MapEntityTest
 
+from geotrek.authent.factories import StructureFactory
 from geotrek.authent.tests import AuthentFixturesTest
 
 
@@ -21,3 +22,23 @@ class CommonTest(AuthentFixturesTest, TranslationResetMixin, MapEntityTest):
 
     def get_bad_data(self):
         return {'topology': 'doh!'}, _('Topology is not valid.')
+
+    def test_structure_is_set(self):
+        if not hasattr(self.model, 'structure'):
+            return
+        self.login()
+        response = self.client.post(self._get_add_url(), self.get_good_data())
+        self.assertEqual(response.status_code, 302)
+        obj = self.model.objects.last()
+        self.assertEqual(obj.structure, self.user.profile.structure)
+
+    def test_structure_is_not_changed(self):
+        if not hasattr(self.model, 'structure'):
+            return
+        self.login()
+        structure = StructureFactory()
+        self.assertNotEqual(structure, self.user.profile.structure)
+        obj = self.modelfactory.create(structure=structure)
+        self.client.post(obj.get_update_url(), self.get_good_data())
+        self.assertEqual(obj.structure, structure)
+